Refactoring product handler

+++ internal/adapters/http/dto/product.go
@@ -11,7 +11,7 @@ type CreateProductRequest struct {
 	Name        string  `json:"name"`
 	Description string  `json:"description"`
 	Price       float64 `json:"price"`
-	CategoryID  uint    `json:"category_id"`
+	CategoryID  int64   `json:"category_id"`
 	ImageURL    string  `json:"image_url"`
 	Stock       int     `json:"stock"`
 }
@@ -48,7 +48,7 @@ type UpdateProductRequest struct {
 	Name        string  `json:"name"`
 	Description string  `json:"description"`
 	Price       float64 `json:"price"`
-	CategoryID  uint    `json:"category_id"`
+	CategoryID  int64   `json:"category_id"`
 	ImageURL    string  `json:"image_url"`
 	Stock       int     `json:"stock"`
 }
