+++ internal/core/ports/category_repository.go
@@ -11,5 +11,5 @@ type CategoryRepository interface {
 	GetCategory(id int64) (*domain.Category, error)
 	GetCategoryByName(name string) *domain.Category
 	GetCategories() []domain.Category
-	ProductsByCategory(id int) ([]domain.Product, error)
+	ProductsByCategory(id int64) ([]domain.Product, error)
 }
